+++ web/src/components/ui/index.ts
@@ -123,3 +123,4 @@ export * from './time-picker';
 export * from './color-picker';
 export * from './gauge';
 export * from './meter';
+export * from './list-section';